修正iPhone 5 上不能正常识别时间戳的问题

+++ EUExConversationView/EUExConversationView/sourceCode/uexCV_TableView/uexCV_TableViewCell.h
@@ -41,7 +41,7 @@ typedef NS_ENUM(NSInteger, uexCV_TableViewCellStatus) {
 @property (nonatomic,strong)UILabel * nameLabel;
 @property (nonatomic,strong)UIView *containerView;
 @property (nonatomic,strong)UILabel * errorLabel;
-@property (nonatomic,assign)NSInteger timestamp;
+@property (nonatomic,assign)long long timestamp;
